+++ src/aux-server/aux-web/aux-player/scene/MapPortalDimensionGroup3D.ts
@@ -21,6 +21,8 @@ import type { AuxBot3DDecoratorFactory } from '../../shared/scene/decorators/Aux
 import type { MapSimulation3D } from './MapSimulation3D';
 
 export class MapPortalDimensionGroup3D extends DimensionGroup3D {
+    name = 'MapPortalDimensionGroup3D';
+    
     simulation3D: MapSimulation3D; // Override base class type.
 
     constructor(