+++ src/aux-server/aux-web/shared/scene/xr/PointerRay3D.ts
@@ -32,6 +32,8 @@ export const PointerRay_DefaultStopDistance: number = 0.25;
 export const PointerRay_DefaultCursorVisible: boolean = false;
 
 export class PointerRay3D extends Object3D {
+    name = 'PointerRay3D';
+    
     /**
      * The ray that this pointer ray 3d object is representing.
      */